Update many dependencies for Matryoshka.
Notably,
- crossbuilding on Scalatl 2.11.8 and 2.12.0
- Scalacheck 1.13 (well, pre-1.14, due to an issue not resolved in 1.13), including new `Cogen` and `Shrink` instances
- crossbuilding `scalacheck` and `test` projects on JVM and JS.
