+++ .travis.yml
@@ -1,6 +1,7 @@
 language: scala
 scala:
   - 2.11.8
+  - 2.12.0
 jdk:
   - oraclejdk8
   - oraclejdk7
