+++ scalacheck/shared/src/main/scala/matryoshka/scalacheck/arbitrary.scala
